requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
opencv-python>=4.5.0
pyyaml>=5.4.0
tqdm>=4.60.0
tensorboard>=2.5.0
pycocotools>=2.0.0
scipy>=1.7.0
ultralytics>=8.0.0
smplx>=0.1.28
einops>=0.4.0

